Address more mailing list feedback

+++ sysdeps/riscv/rv32/rvd/s_lround.c
@@ -18,11 +18,12 @@
 
 #include <math.h>
 #include <libm-alias-double.h>
+#include <stdint.h>
 
 long int
 __lround (double x)
 {
-  long int res;
+  int32_t res;
   asm ("fcvt.w.d %0, %1, rmm" : "=r" (res) : "f" (x));
   return res;
 }
