+++ sysdeps/riscv/rv64/rvd/s_llrint.c
@@ -17,13 +17,14 @@
    <http://www.gnu.org/licenses/>.  */
 
 #include <libm-alias-double.h>
+#include <stdint.h>
 
 long long int
 __llrint (double x)
 {
-  long long int res;
+  int64_t res;
   asm ("fcvt.l.d %0, %1" : "=r" (res) : "f" (x));
   return res;
 }
 
-libm_alias_double(__llrint, llrint)
+libm_alias_double (__llrint, llrint)
